+++ examples/bad_models.py
@@ -2,6 +2,7 @@ from django.db import models
 from django.contrib.postgres.fields import ArrayField
 from django.contrib.postgres.search import SearchVector
 from django.contrib.postgres.aggregates import ArrayAgg
+from django.db.models import Count, Subquery
 
 
 class Widget(models.Model):
@@ -17,3 +18,10 @@ class Widget(models.Model):
 
     def unique_names(self):
         return Widget.objects.order_by("name").distinct("name")
+
+    def with_latest_note(self):
+        latest_note = Note.objects.filter(widget=models.OuterRef("pk")).values("text")[:1]
+        return Widget.objects.annotate(
+            part_count=Count("parts", distinct=True),
+            latest_note=Subquery(latest_note),
+        )